+++ src/schemathesis/__init__.py
@@ -2,7 +2,7 @@ from . import hooks
 from ._hypothesis import init_default_strategies, register_string_format
 from .cli import register_check
 from .constants import __version__
-from .loaders import Parametrizer, from_dict, from_file, from_path, from_pytest_fixture, from_uri, from_wsgi
+from .loaders import from_dict, from_file, from_path, from_pytest_fixture, from_uri, from_wsgi
 from .models import Case
 
 init_default_strategies()